+++ src/Credfeto.ChangeLog.Tests/ChangeLogLinterTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Credfeto.ChangeLog.Models;
 using FunFair.Test.Common;
 using Xunit;