+++ src/Credfeto.ChangeLog/Extensions/ChangeLogHeadingExtensions.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Credfeto.ChangeLog.Helpers;
+namespace Credfeto.ChangeLog.Extensions;
 
 internal static class ChangeLogHeadingExtensions
 {