+++ src/Credfeto.ChangeLog/ServiceCollectionExtensions.cs
@@ -1,3 +1,4 @@
+using Credfeto.ChangeLog.Services;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Credfeto.ChangeLog;