+++ src/Credfeto.ChangeLog/Services/ChangeLogReaderService.cs
@@ -1,9 +1,10 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Credfeto.ChangeLog;
 using Credfeto.ChangeLog.Helpers;
 
-namespace Credfeto.ChangeLog;
+namespace Credfeto.ChangeLog.Services;
 
 internal sealed class ChangeLogReaderService : IChangeLogReader
 {