+++ src/Credfeto.ChangeLog/Services/ChangeLogUpdaterService.cs
@@ -1,7 +1,8 @@
 using System.Threading;
 using System.Threading.Tasks;
+using Credfeto.ChangeLog;
 
-namespace Credfeto.ChangeLog;
+namespace Credfeto.ChangeLog.Services;
 
 internal sealed class ChangeLogUpdaterService : IChangeLogUpdater
 {