+++ src/Credfeto.ChangeLog/Services/IChangeLogUpdater.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Credfeto.ChangeLog;
+namespace Credfeto.ChangeLog.Services;
 
 public interface IChangeLogUpdater
 {